Fixed silly expected result issue
