+++ src/app/guide-list/page.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 export default async function GuideListPage() {
   const session = await auth();
 
-  const roll = session!.user.role;
+  const roll = session ? session.user.role : "";
   //get all guides
   const guides = await prisma.blogPost.findMany(); //page this
 
@@ -13,7 +13,10 @@ export default async function GuideListPage() {
 
   const editButton = (id: string) => {
     return (
-      <Link className="bg-neutral-600 h-full px-1 place-content-center" href={`/curator/guide/edit/${id}`}>
+      <Link
+        className="h-full place-content-center bg-neutral-600 px-1"
+        href={`/curator/guide/edit/${id}`}
+      >
         <svg
           xmlns="http://www.w3.org/2000/svg"
           width="24"
@@ -33,16 +36,19 @@ export default async function GuideListPage() {
   };
 
   return (
-    <div className="flex flex-col h-full w-full pl-16 p-2 gap-4">
+    <div className="flex h-full w-full flex-col gap-4 p-2 pl-16">
       <h1 className="text-5xl font-bold">Guides</h1>
-      <div className="grid h-full w-2xl gap-8 grid-cols-2">
+      <div className="grid h-full w-2xl grid-cols-2 gap-8">
         {guides.map((post, index) => {
           return (
             <div
-              className="flex h-16 w-full overflow-clip flex-row place-content-between place-items-center rounded-lg bg-neutral-800"
+              className="flex h-16 w-full flex-row place-content-between place-items-center overflow-clip rounded-lg bg-neutral-800"
               key={index}
             >
-              <Link className="w-full px-2 line-clamp-2" href={`/guide/${post.id}`}>
+              <Link
+                className="line-clamp-2 w-full px-2"
+                href={`/guide/${post.id}`}
+              >
                 {post.title}
               </Link>
               {roll === "CURATOR" || roll === "ADMIN"
