+++ src/auth.ts
@@ -1,5 +1,36 @@
 import NextAuth from "next-auth";
-import authConfig from "./auth.config";
+import { PrismaAdapter } from "@auth/prisma-adapter";
+import { PoolConfig } from "@neondatabase/serverless";
+import { PrismaNeon } from "@prisma/adapter-neon";
+import { Provider } from "next-auth/providers";
+import GitHub from "next-auth/providers/github";
+import { PrismaClient } from "../generated/prisma";
+
+const neon: PoolConfig = {
+  connectionString: process.env.AUTH_POSTGRES_PRISMA_URL,
+};
+const adapter = new PrismaNeon(neon);
+const prisma = new PrismaClient({ adapter });
+
+const providers: Provider[] = [
+  GitHub({
+    clientId: process.env.AUTH_GITHUB_ID,
+    clientSecret: process.env.AUTH_GITHUB_SECRET,
+    allowDangerousEmailAccountLinking: true,
+  }),
+];
+
+export const providerMap = providers
+  .map((provider) => {
+    if (typeof provider === "function") {
+      const providerData = provider();
+      return { id: providerData.id, name: providerData.name };
+    } else {
+      return { id: provider.id, name: provider.name };
+    }
+  })
+  .filter((provider) => provider.id !== "credentials");
+
 
 export const {
   handlers,
@@ -8,7 +39,9 @@ export const {
   signOut,
   unstable_update: update,
 } = NextAuth({
-  ...authConfig,
+  adapter: PrismaAdapter(prisma),
+  providers: providers,
+
   // pages: { signIn: "/signin" },
   session: { strategy: "database" },
   callbacks: {
